moved the daysSinceLastExercise to `./exercises/utils`. It logically fits better there for me.
